pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pyaptamer"
version = "0.1.0"
description = "Aptamer toolkit for PDB/mmCIF structure handling"
authors = [{name = "Avinab Neogy", email = "avinabneogy.bits@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "biopython>=1.80",
    "requests>=2.28.0", 
    "numpy>=1.21.0",
    "gemmi>=0.6.3"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["pyaptamer*"]
